Separated signal handlers in different modules.
